backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
alembic==1.13.2
pydantic==2.9.2
pydantic-settings==2.6.1
email-validator==2.1.0.post1
python-dotenv==1.0.1
loguru==0.7.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
typing-extensions==4.12.2
aiosqlite==0.20.0
websockets==13.1
sentence-transformers==3.0.1
scikit-learn==1.5.2
joblib==1.4.2
httpx==0.27.2
google-generativeai>=0.4.1
openai>=1.14.0
huggingface-hub>=0.21.4
aiohttp>=3.9.0
